[af-207] fix: Product alcohol columnDefinition 수정

+++ src/main/java/com/drunkenlion/alcoholfriday/domain/product/entity/Product.java
@@ -37,7 +37,7 @@ public class Product extends BaseEntity {
     private Long quantity;
 
     @Comment("술 도수")
-    @Column(name = "alcohol", columnDefinition = "BIGINT")
+    @Column(name = "alcohol", columnDefinition = "DOUBLE")
     private Double alcohol;
 
     @Comment("제품 재료")
